Extract Model Metadata Info To Separate Module
